+++ SchoolManagement/Controllers/StudentsController.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using SchoolManagement.Helpers.DTOs;
@@ -7,19 +8,23 @@ using SchoolManagement.Interfaces;
 
 namespace SchoolManagement.Controllers
 {
+    
     [ApiController]
     [Route("api/[controller]")]
+    //[Authorize]
     public class StudentsController : ControllerBase
     {
         private readonly IStudentRepository _studentRepository;
         private readonly ILogger<StudentsController> _logger;
         private readonly IMapper _mapper;
+        private readonly IAuthenticateUser _tokenService;
 
-        public StudentsController(IStudentRepository studentRepository, ILogger<StudentsController> logger, IMapper mapper)
+        public StudentsController(IStudentRepository studentRepository, ILogger<StudentsController> logger, IMapper mapper, IAuthenticateUser tokenService)
         {
             _studentRepository = studentRepository;
             _logger = logger;
-            _mapper = mapper;   
+            _mapper = mapper;
+            _tokenService = tokenService;
         }
 
         // Endpoint to create a new student
@@ -34,6 +39,22 @@ namespace SchoolManagement.Controllers
 
             try
             {
+                var token = Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+
+                if (string.IsNullOrEmpty(token))
+                {
+                    _logger.LogWarning("Token not found in the request header.");
+                    return BadRequest(new BaseResponse { ResponseCode = "01", ResponseMessage = "Token not found" });
+                }
+
+                var principal = _tokenService.ValidateToken(token);
+
+                if (principal == null)
+                {
+                    _logger.LogWarning("Invalid token.");
+                    return BadRequest(new BaseResponse { ResponseCode = "02", ResponseMessage = "Invalid token" });
+                }
+
                 var ran = new Random();
 
                 var studentModel = _mapper.Map<Student>(student);
@@ -68,6 +89,24 @@ namespace SchoolManagement.Controllers
 
             try
             {
+                //First, authenticate the JWT token supplied by the client
+                var token = Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+
+                if (string.IsNullOrEmpty(token))
+                {
+                    _logger.LogWarning("Token not found in the request header.");
+                    return BadRequest(new BaseResponse { ResponseCode = "01", ResponseMessage = "Token not found" });
+                }
+
+                var principal = _tokenService.ValidateToken(token);
+
+                if (principal == null)
+                {
+                    _logger.LogWarning("Invalid token.");
+                    return BadRequest(new BaseResponse { ResponseCode = "02", ResponseMessage = "Invalid token" });
+                }
+
+                //Jwt token is valid, proceed to complete the request
                 var students = await _studentRepository.GetAllAsync();
                 _logger.LogInformation("Students retrieved successfully.");
                 return StatusCode(StatusCodes.Status200OK, new CreateStudentResponse { StudentInfo = students, ResponseCode = "00", ResponseMessage = "Successful Fetch" });
@@ -84,12 +123,30 @@ namespace SchoolManagement.Controllers
         [ProducesResponseType(typeof(BaseResponse), StatusCodes.Status404NotFound)]
         [ProducesResponseType(typeof(BaseResponse), StatusCodes.Status500InternalServerError)]
         [HttpGet("{studentId}")]
-        public async Task<ActionResult<Student>> GetStudent(int studentId)
+        public async Task<ActionResult<Student>> GetStudent(string studentId)
         {
             _logger.LogInformation($"Request received to get student with ID: {studentId}");
 
             try
             {
+                //First, authenticate the JWT token supplied by the client
+                var token = Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+
+                if (string.IsNullOrEmpty(token))
+                {
+                    _logger.LogWarning("Token not found in the request header.");
+                    return BadRequest(new BaseResponse { ResponseCode = "01", ResponseMessage = "Token not found" });
+                }
+
+                var principal = _tokenService.ValidateToken(token);
+
+                if (principal == null)
+                {
+                    _logger.LogWarning("Invalid token.");
+                    return BadRequest(new BaseResponse { ResponseCode = "02", ResponseMessage = "Invalid token" });
+                }
+
+                //Jwt token is valid, proceed to complete the request
                 var student = await _studentRepository.GetByIdAsync(studentId);
                 if (student == null)
                 {
@@ -118,6 +175,24 @@ namespace SchoolManagement.Controllers
 
             try
             {
+                //First, authenticate the JWT token supplied by the client
+                var token = Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+
+                if (string.IsNullOrEmpty(token))
+                {
+                    _logger.LogWarning("Token not found in the request header.");
+                    return BadRequest(new BaseResponse { ResponseCode = "01", ResponseMessage = "Token not found" });
+                }
+
+                var principal = _tokenService.ValidateToken(token);
+
+                if (principal == null)
+                {
+                    _logger.LogWarning("Invalid token.");
+                    return BadRequest(new BaseResponse { ResponseCode = "02", ResponseMessage = "Invalid token" });
+                }
+
+                //Jwt token is valid, proceed to complete the request
                 var studentModel = _mapper.Map<Student>(student);
 
                 var successful = await _studentRepository.UpdateAsync(studentModel);
@@ -145,12 +220,30 @@ namespace SchoolManagement.Controllers
         [ProducesResponseType(typeof(BaseResponse), StatusCodes.Status400BadRequest)]
         [ProducesResponseType(typeof(BaseResponse), StatusCodes.Status500InternalServerError)]
         [HttpDelete("{studentId}")]
-        public async Task<IActionResult> DeleteStudent(int studentId)
+        public async Task<IActionResult> DeleteStudent(string studentId)
         {
             _logger.LogInformation($"Request received to delete student with ID: {studentId}");
 
             try
             {
+                //First, authenticate the JWT token supplied by the client
+                var token = Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+
+                if (string.IsNullOrEmpty(token))
+                {
+                    _logger.LogWarning("Token not found in the request header.");
+                    return BadRequest(new BaseResponse { ResponseCode = "01", ResponseMessage = "Token not found" });
+                }
+
+                var principal = _tokenService.ValidateToken(token);
+
+                if (principal == null)
+                {
+                    _logger.LogWarning("Invalid token.");
+                    return BadRequest(new BaseResponse { ResponseCode = "02", ResponseMessage = "Invalid token" });
+                }
+
+                //Jwt token is valid, proceed to complete the request
                 var successful = await _studentRepository.DeleteAsync(studentId);
 
                 if (successful)
